New option for vitality FIxed some event engine null pointers Correct wrong structure for character_reco_bonus table Fixed pet consume food even if you are mounted Fixed skill exploit if you change fast target Retail like veil skill Fixed some random errors Fixed some items does not consume
Core
compile and copy paste core.jar and event-engine.jar
check Rates.ini changes

Data
Delete scripts folder and paste form this rev
Delete xml folder and paste from this rev

for live servers Execute also
Update_(12_02_2016).sql located in tools/sql/game/updates/

+++ L2J_SunriseProject_Core/java/l2r/gameserver/model/actor/instance/L2PetInstance.java
@@ -1155,7 +1155,10 @@ public class L2PetInstance extends L2Summon
 		{
 			if (getInventory() != null)
 			{
-				getInventory().deleteMe();
+				if (!owner.isMounted())
+				{
+					getInventory().deleteMe();
+				}
 			}
 			L2World.getInstance().removePet(owner.getObjectId());
 		}
